fix: fixed debug variable
Debug variable can now be set to 'false' on the
production deploy.

This fixes issue #1.

fix: downgraded pip version

Due to pip messing with the psycopg-binary package,
we had to downgrade pip in order to continue working.

More about this issue in #7.

fix: debug variable changed to false in production

This partially fixes issue #1.

fix: change settings.py to base.py

This could be causing problems due to not following name convention on settings directory.