ci: execute behavior tests on every branch
Before this commit, the workflow was only executing on develop and main,
which caused some commits like for example 1e2bb3945a09561a1f2ebe3f4afb99e015c23b97
to fail, but not because of the code of this commit, but on other commit
merged before 1e2bb3945a09561a1f2ebe3f4afb99e015c23b97 which didn't pass
the tests!

Now every person that creates a branch and pushes to it will know if it
is in a unstable version (didn't pass all the tests) or a stable one.